#84: Move ConfigUtils from shared to api module
